- Removed overhead - Liquid objects can now be used within the file_exists tag - Added a detailed usage description
