--- _layouts/base.html
@@ -1,9 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <title>Test</title>
-</head>
-<body>
-    {{ content }}
-</body>
-</html>
