--- index.html
@@ -1,13 +1,0 @@
----
-  layout: base
----
-
-<p>Available: {% file_exists images/sample.png %}</p>
-<p>Missing: {% file_exists images/missing.png %}</p>
-
-<!-- Now it would be cool if we could wrap this in if, this would me to where I need to be -->
-
-<!--
-<p>Available: {% if file_exists images/sample.png %}true{% endif %}</p>
-<p>Missing: {% if file_exists images/missing.png %}false{% endif %}</p>
--->
